+++ changelog.md
@@ -1,2 +1,2 @@
 ### Tesseract 1.0.37
-- Initial release of Tesseract for Minecraft 1.21.5
+- Initial release of Tesseract for Minecraft 1.21.6